third_party/blink/web_tests/platform/generic/accessibility/aria-readonly-expected.txt

     

This tests that the AXValue property is correctly reported for native and non-native text boxes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ariaTextBoxIsWritable is true
PASS ariaReadOnlyAriaTextBoxIsWritable is false
PASS ariaReadOnlyTextFieldIsWritable is false
PASS ariaNonReadOnlyTextFieldIsWritable is true
PASS htmlReadOnlyTextFieldIsWritable is false
PASS textFieldIsWritable is true
PASS htmlReadOnlyTextAreaIsWritable is false
PASS textAreaIsWritable is true
PASS successfullyParsed is true

TEST COMPLETE
